Add Autoincrement ID, change SearchItem -> Search function

+++ src/Person.cs
@@ -5,7 +5,7 @@ namespace management_system
 {
     public class Person
     {
-        protected int Id { get; init; }
+        protected long Id { get; init; }
         protected string Name { get; init; }
         protected string Surname { get; init; }
         protected string Email { get; init; }
